Factor out tracing() (i.e. a "context manager").
